requirements.txt
django==1.5.1
south==0.8.1
django-tastypie==0.9.15
django-admin-tools==0.5.1
feedparser
PIL
requests
django-extensions
pygraphviz
scrapy==0.16.4
psycopg2==2.5
